Unlock logind sessions on resume
Most desktop environments lock the screen before suspending.
Thus, on resume, the user has to enter two passwords: one
for unlocking the encrypted root device, and a second one
to unlock the screen. This is annoying and unneeded as
knowing the passphrase for the root device is sufficient
proof of authenticity.

This patch unlocks the screen automatically on successful
resume. As we do not know the ID of the user session, we
unlock all sessions (on a one-user laptop, there is usually
only one active session).