Move sync syscall just before luksSuspend
This avoids that a following sync syscall in in pm_suspend()
(called when we echo mem > /sys/power/state) hangs forever
in at least one case.

I guess that one case is the following: if tracking access time
is enabled on an ext4 file system, then executing chroot in
arch-luks-suspend will require I/O in the sync syscall in
pm_suspend().